+++ src/pytestarch/eval_structure/evaluable_graph.py
@@ -1,21 +1,44 @@
 """Base class for different graph implementations of an evaluable structure. Delegates direct access to graph nodes
 and edges to its subclasses in a template pattern.
 """
-from typing import Any, List, Optional, Set, Tuple
+from itertools import product
+from typing import Any, List, Set, Union
 
-from pytestarch.eval_structure.eval_structure_types import EvaluableArchitecture, Module
-from pytestarch.eval_structure.graph import Graph, Node
+from pytestarch.eval_structure.evaluable_architecture import (
+    EvaluableArchitecture,
+    LaxDependenciesByBaseModule,
+    Module,
+    StrictDependenciesByBaseModules,
+    StrictDependency,
+)
+from pytestarch.eval_structure.evaluable_structures import AbstractGraph, AbstractNode
 
 
 class EvaluableArchitectureGraph(EvaluableArchitecture):
     """Abstract implementation of an evaluable object that is based on a graph structure."""
 
-    def __init__(self, graph: Graph) -> None:
+    def __init__(self, graph: AbstractGraph) -> None:
         self._graph = graph
 
-    def get_dependency(
+    def get_dependencies(
+        self,
+        dependents: Union[Module, List[Module]],
+        dependent_upons: Union[Module, List[Module]],
+    ) -> StrictDependenciesByBaseModules:
+        result = {}
+
+        dependents = self._listify(dependents)
+        dependent_upons = self._listify(dependent_upons)
+
+        for dependent, dependent_upon in product(dependents, dependent_upons):
+            dependency = self._get_dependency_between_modules(dependent, dependent_upon)
+            result[(dependent, dependent_upon)] = dependency
+
+        return result
+
+    def _get_dependency_between_modules(
         self, dependent: Module, dependent_upon: Module
-    ) -> Optional[Tuple[str, str]]:
+    ) -> List[StrictDependency]:
         dependent_node = self._get_node(dependent)
         dependent_upon_nodes = self._get_all_submodules_of(dependent_upon)
 
@@ -23,6 +46,8 @@ class EvaluableArchitectureGraph(EvaluableArchitecture):
 
         nodes_to_check = [dependent_node]
         checked_nodes = set()
+
+        dependencies = []
 
         while nodes_to_check:
             node = nodes_to_check.pop()
@@ -43,18 +68,39 @@ class EvaluableArchitectureGraph(EvaluableArchitecture):
                     and node not in nodes_to_exclude
                     and child not in nodes_to_exclude
                 ):
-                    return node, child
+                    dependencies.append(tuple(self._to_modules([node, child])))
+        return dependencies  # type: ignore
 
-        return None
-
-    def _get_node(self, dependent: Module) -> Node:
+    def _get_node(self, dependent: Module) -> AbstractNode:
         return dependent.name or dependent.parent_module
 
-    def any_dependency_to_module_other_than(
-        self, dependent: Module, dependent_upon: Module
-    ) -> List[Module]:
-        # submodules of the dependent upon module do not count as an import that is not the dependent upon module
-        nodes_to_exclude = self._get_all_submodules_of(dependent_upon)
+    def any_dependencies_to_modules_other_than(
+        self,
+        dependents: Union[Module, List[Module]],
+        dependent_upons: Union[Module, List[Module]],
+    ) -> LaxDependenciesByBaseModule:
+        result = {}
+
+        dependents = self._listify(dependents)
+        dependent_upons = self._listify(dependent_upons)
+
+        for dependent in dependents:
+            dependencies = self._any_dependency_to_module_other_than(
+                dependent, dependent_upons
+            )
+            result[dependent] = dependencies
+
+        return result
+
+    def _listify(self, module: Union[Module, List[Module]]) -> List[Module]:
+        return [module] if not isinstance(module, list) else module
+
+    def _any_dependency_to_module_other_than(  # noqa: C901
+        self, dependent: Module, dependent_upons: List[Module]
+    ) -> List[StrictDependency]:
+        nodes_to_exclude = set()
+        for dependent_upon in dependent_upons:
+            nodes_to_exclude.update(self._get_all_submodules_of(dependent_upon))
 
         nodes_fulfilling_criteria = []
         # should submodules of the dependent module import each other, this does not count as a dependency
@@ -65,18 +111,16 @@ class EvaluableArchitectureGraph(EvaluableArchitecture):
             # true submodules should be considered
             nodes_to_exclude.add(dependent.parent_module)
 
-        if dependent_upon.parent_module is not None:
-            # if there is a dependency to the parent module of dependent upon, this counts, as only dependencies to
-            # the true submodules are excluded
-            nodes_to_exclude.remove(dependent_upon.parent_module)
+        for dependent_upon in dependent_upons:
+            if dependent_upon.parent_module is not None:
+                # if there is a dependency to the parent module of dependent upon, this counts, as only dependencies to
+                # the true submodules are excluded
+                nodes_to_exclude.remove(dependent_upon.parent_module)
 
         nodes_to_check = list(nodes_that_do_not_fulfill_criterion)
         checked_nodes = set()
 
         while nodes_to_check:
-            if nodes_fulfilling_criteria:
-                break
-
             node = nodes_to_check.pop()
 
             if node in checked_nodes:
@@ -96,17 +140,40 @@ class EvaluableArchitectureGraph(EvaluableArchitecture):
                         child not in nodes_to_exclude
                         and child not in nodes_that_do_not_fulfill_criterion
                     ):
-                        nodes_fulfilling_criteria.append(child)
+                        nodes_fulfilling_criteria.append(
+                            tuple(self._to_modules([node, child]))
+                        )
+                    else:
+                        nodes_to_check.append(child)
 
-                    nodes_to_check.append(child)
+        return nodes_fulfilling_criteria  # type: ignore
 
-        return self._to_modules(nodes_fulfilling_criteria)
+    def any_other_dependencies_to_modules_than(
+        self,
+        dependents: Union[Module, List[Module]],
+        dependent_upons: Union[Module, List[Module]],
+    ) -> LaxDependenciesByBaseModule:
+        result = {}
 
-    def any_other_dependency_to_module_than(
-        self, dependent: Module, dependent_upon: Module
-    ) -> List[Module]:
+        dependents = self._listify(dependents)
+        dependent_upons = self._listify(dependent_upons)
+
+        for dependent_upon in dependent_upons:
+            dependencies = self._any_other_dependency_to_module_than(
+                dependents, dependent_upon
+            )
+            result[dependent_upon] = dependencies
+
+        return result
+
+    def _any_other_dependency_to_module_than(  # noqa: C901
+        self, dependents: List[Module], dependent_upon: Module
+    ) -> List[StrictDependency]:
+        # submodules of the dependent upon module do not count as an import that is not the dependent upon module
         # submodules of and including the dependent do not count as allowed imports
-        nodes_to_exclude = self._get_all_submodules_of(dependent)
+        nodes_to_exclude = set()
+        for dependent in dependents:
+            nodes_to_exclude.update(self._get_all_submodules_of(dependent))
 
         nodes_fulfilling_criteria = []
         # submodules of the dependent upon do not count as allowed imports if they should import each other
@@ -123,9 +190,10 @@ class EvaluableArchitectureGraph(EvaluableArchitecture):
 
         nodes_to_check = list(nodes_that_count_as_not_fulfilling_criterion)
 
-        if dependent.parent_module is not None:
-            # if the dependent module is defined via a parent module, this parent module counts as an allowed import
-            nodes_to_exclude.remove(dependent.parent_module)
+        for dependent in dependents:
+            if dependent.parent_module is not None:
+                # if the dependent module is defined via a parent module, this parent module counts as an allowed import
+                nodes_to_exclude.remove(dependent.parent_module)
 
         checked_nodes = set()
 
@@ -146,17 +214,19 @@ class EvaluableArchitectureGraph(EvaluableArchitecture):
                         parent not in nodes_to_exclude
                         and parent not in nodes_that_count_as_not_fulfilling_criterion
                     ):
-                        nodes_fulfilling_criteria.append(parent)
+                        nodes_fulfilling_criteria.append(
+                            tuple(self._to_modules([parent, node]))
+                        )
 
                     if parent not in nodes_to_exclude:
                         nodes_to_check.append(parent)
 
-        return self._to_modules(nodes_fulfilling_criteria)
+        return nodes_fulfilling_criteria  # type: ignore
 
-    def _to_modules(self, nodes: List[Node]) -> List[Module]:
+    def _to_modules(self, nodes: List[AbstractNode]) -> List[Module]:
         return list(map(lambda node: Module(name=node), nodes))
 
-    def _get_all_submodules_of(self, module: Module) -> Set[Node]:
+    def _get_all_submodules_of(self, module: Module) -> Set[AbstractNode]:
         """Returns all submodules of a given module.
 
         Args:
@@ -192,6 +262,6 @@ class EvaluableArchitectureGraph(EvaluableArchitecture):
     def visualize(self, **kwargs: Any) -> None:
         self._graph.draw(**kwargs)
 
-    def _get_excluded_nodes(self, modules: List[Module]) -> List[Node]:
+    def _get_excluded_nodes(self, modules: List[Module]) -> List[AbstractNode]:
         result = [module.parent_module for module in modules]
         return [node for node in result if node is not None]