+++ src/pytestarch/query_language/base_language.py
@@ -1,31 +1,21 @@
 from __future__ import annotations
 
+import os
 from abc import ABC, abstractmethod
-from collections import defaultdict
-from typing import Generic, Optional, Type, TypeVar
+from typing import Callable, Generic, List, Optional, Type, TypeVar, Union
 
-from pytestarch.eval_structure.eval_structure_types import EvaluableArchitecture, Module
-from pytestarch.eval_structure.graph import Node
-from pytestarch.query_language.exceptions import ImproperlyConfigured
+from pytestarch.eval_structure.evaluable_architecture import (
+    EvaluableArchitecture,
+    Module,
+)
+from pytestarch.exceptions import ImproperlyConfigured
+from pytestarch.query_language.message_generator import RuleViolationMessageGenerator
 from pytestarch.query_language.rule_matcher import (
     BehaviorRequirement,
     DefaultRuleMatcher,
     ModuleRequirement,
     RuleMatcher,
     RuleViolations,
-)
-
-# (Import, negated, singular)
-PREFIX_MAPPING = defaultdict(str)
-PREFIX_MAPPING.update(
-    {
-        (False, False, True): "is ",
-        (False, True, True): "is not ",
-        (True, True, True): "does not ",
-        (False, False, False): "are ",
-        (False, True, False): "are not ",
-        (True, True, False): "do not ",
-    }
 )
 
 
@@ -71,11 +61,13 @@ class ModuleSpecification(Generic[ModuleSpecificationSuccessor], ABC):
     """Offers functionality to specify detail information about Rule Subjects or Objects."""
 
     @abstractmethod
-    def are_sub_modules_of(self, modules: Node) -> ModuleSpecificationSuccessor:
+    def are_sub_modules_of(
+        self, modules: Union[str, List[str]]
+    ) -> ModuleSpecificationSuccessor:
         pass
 
     @abstractmethod
-    def are_named(self, name: str) -> ModuleSpecificationSuccessor:
+    def are_named(self, names: Union[str, List[str]]) -> ModuleSpecificationSuccessor:
         pass
 
 
@@ -133,7 +125,7 @@ class Rule(
         self._rule_matcher_class = rule_matcher_class
 
         self._module_to_check: Optional[Module] = None
-        self._module_to_check_against: Optional[Module] = None
+        self._modules_to_check_against: Optional[List[Module]] = None
 
         self._should = False
         self._should_only = False
@@ -153,31 +145,38 @@ class Rule(
         self._modules_to_check_to_be_specified_next = True
         return self
 
-    def are_sub_modules_of(self, module: Node) -> BehaviorSpecification:
+    def are_sub_modules_of(
+        self, modules: Union[str, List[str]]
+    ) -> BehaviorSpecification:
+        self._set_modules(modules, lambda name: Module(parent_module=name))
+        return self
+
+    def are_named(self, names: Union[str, List[str]]) -> BehaviorSpecification:
+        self._set_modules(names, lambda name: Module(name=name))
+        return self
+
+    def _set_modules(
+        self,
+        module_names: Union[str, List[str]],
+        create_module_fn: Callable[[str], Module],
+    ) -> None:
         if self._modules_to_check_to_be_specified_next is None:
             raise ImproperlyConfigured("Specify a RuleSubject or RuleObject first.")
 
-        module = Module(parent_module=module)
+        if self._modules_to_check_to_be_specified_next and isinstance(
+            module_names, list
+        ):
+            raise ImproperlyConfigured("Only rule subjects can be specified in batch.")
+
+        if not self._modules_to_check_to_be_specified_next and isinstance(
+            module_names, str
+        ):
+            module_names = [module_names]
 
         if self._modules_to_check_to_be_specified_next:
-            self._module_to_check = module
+            self._module_to_check = create_module_fn(module_names)  # type: ignore
         else:
-            self._module_to_check_against = module
-
-        return self
-
-    def are_named(self, name: str) -> BehaviorSpecification:
-        if self._modules_to_check_to_be_specified_next is None:
-            raise ImproperlyConfigured("Specify a RuleSubject or RuleObject first.")
-
-        module = Module(name=name)
-
-        if self._modules_to_check_to_be_specified_next:
-            self._module_to_check = module
-        else:
-            self._module_to_check_against = module
-
-        return self
+            self._modules_to_check_against = [create_module_fn(n) for n in module_names]
 
     def should(self) -> DependencySpecification:
         self._should = True
@@ -216,8 +215,15 @@ class Rule(
     def assert_applies(self, evaluable: EvaluableArchitecture) -> None:
         self._assert_required_configuration_present()
 
+        matcher = self._prepare_rule_matcher()
+        rule_violations = matcher.match(evaluable)
+
+        if rule_violations:
+            raise AssertionError(self._create_rule_violation_message(rule_violations))
+
+    def _prepare_rule_matcher(self) -> RuleMatcher:
         module_requirement = ModuleRequirement(
-            self._module_to_check, self._module_to_check_against, self._import
+            self._module_to_check, self._modules_to_check_against, self._import
         )
         behavior_requirement = BehaviorRequirement(
             self._should,
@@ -226,23 +232,23 @@ class Rule(
             self._except_present,
         )
 
-        rule_violations = self._rule_matcher_class(
-            module_requirement, behavior_requirement
-        ).match(evaluable)
-
-        if rule_violations:
-            raise AssertionError(self._create_rule_violation_message(rule_violations))
+        return self._rule_matcher_class(module_requirement, behavior_requirement)
 
     def __str__(self) -> str:
+        self._assert_required_configuration_present()
+
         method_name = f'{"should" if self._should else "should only" if self._should_only else "should not"}'
 
+        subject_prefix = (
+            "Sub modules of " if self._module_to_check.parent_module is not None else ""
+        )
         return (
-            f"{self._add_rule_subject_plural_marker_if_needed(self._get_module_name(self._module_to_check))} "
+            f'{subject_prefix}"{self._get_module_name(self._module_to_check)}" '
             f"{method_name} "
             f'{"import" if self._import else "be imported by"} '
             f'{"modules except " if self._except_present else ""}'
-            f'modules that are {"named" if self._module_to_check_against.name is not None else "sub modules of"} '
-            f'"{self._get_module_name(self._module_to_check_against)}".'
+            f'modules that are {"named" if self._modules_to_check_against[0].name is not None else "sub modules of"} '
+            f'"{", ".join(map(lambda module: self._get_module_name(module), self._modules_to_check_against))}".'
         )
 
     def _get_module_name(self, module: Module) -> str:
@@ -252,7 +258,7 @@ class Rule(
         behavior_missing = not any([self._should, self._should_only, self._should_not])
         dependency_missing = self._import is None
         subject_missing = not self._module_to_check
-        object_missing = not self._module_to_check_against
+        object_missing = not self._modules_to_check_against
 
         if any([behavior_missing, dependency_missing, subject_missing, object_missing]):
             subject_message = self._name_or_empty(subject_missing, RuleSubject)
@@ -282,128 +288,12 @@ class Rule(
         return f"a {clz.__name__}" if empty else ""
 
     def _create_rule_violation_message(self, rule_violations: RuleViolations) -> str:
-        rule_subject = self._create_rule_subject_for_rule_violation_message(
+        message_generator = RuleViolationMessageGenerator(
+            self._module_to_check,
+            self._modules_to_check_against,
+            self._import,
+        )
+        single_violation_messages = message_generator.create_rule_violation_messages(
             rule_violations
         )
-        verb = self._create_verb_for_rule_violation_message(rule_violations)
-        rule_object = self._create_rule_object_for_rule_violation_message(
-            rule_violations
-        )
-        return f"{rule_subject} {verb}{rule_object}."
-
-    def _create_rule_subject_for_rule_violation_message(
-        self, rule_violations: RuleViolations
-    ) -> str:
-        module_name = self._get_module_name(self._module_to_check)
-
-        if self._forbidden_import(rule_violations):
-            if rule_violations.strict_dependency is not None:
-                module_name = self._get_subject_of_strict_dependency(rule_violations)
-
-                # no plural marker needed, as module is always singular
-                return f'"{module_name}"'
-
-        return self._add_rule_subject_plural_marker_if_needed(module_name)
-
-    def _get_subject_of_strict_dependency(self, rule_violations: RuleViolations) -> str:
-        # strict dependency is always reported in format (importer, importee)
-        # if rule is of format A ... imports B, the dependency will be (A, B)
-        idx = int(not self._import)
-        return rule_violations.strict_dependency[idx]
-
-    def _get_object_of_strict_dependency(self, rule_violations: RuleViolations) -> str:
-        # strict dependency is always reported in format (importer, importee)
-        # if rule is of format A ... be imported by B, the dependency will be (B, A)
-        # but in our rule, B is the object
-        idx = int(self._import)
-        return rule_violations.strict_dependency[idx]
-
-    def _add_rule_subject_plural_marker_if_needed(self, module_name: str) -> str:
-        prefix = ""
-        if not self._rule_subject_singular:
-            prefix = "Sub modules of "
-        return f'{prefix}"{module_name}"'
-
-    def _create_verb_for_rule_violation_message(
-        self, rule_violations: RuleViolations
-    ) -> str:
-        base_verb = "import" if self._import else "imported by"
-
-        suffix = " "
-
-        if rule_violations.should_violated:
-            negated = True
-
-        elif (
-            rule_violations.should_except_violated
-            or rule_violations.should_only_except_violated_by_no_import
-        ):
-            negated = True
-
-        elif (
-            rule_violations.should_only_except_violated_by_forbidden_import
-            or rule_violations.should_not_violated
-            or rule_violations.should_only_violated
-            or rule_violations.should_not_except_violated
-        ):
-
-            if self._import and (
-                self._rule_subject_singular or (self._forbidden_import(rule_violations))
-            ):
-                suffix = "s "
-
-            negated = False
-        else:
-            raise Exception("Unknown rule violation detected.")
-
-        prefix = PREFIX_MAPPING[
-            (
-                self._import,
-                negated,
-                self._rule_subject_singular
-                or (self._forbidden_import(rule_violations)),
-            )
-        ]
-
-        return f"{prefix}{base_verb}{suffix}"
-
-    def _forbidden_import(self, rule_violations: RuleViolations) -> bool:
-        return (
-            rule_violations.should_not_violated
-            or rule_violations.should_only_except_violated_by_forbidden_import
-        )
-
-    @property
-    def _rule_subject_singular(self) -> bool:
-        return self._module_to_check.name is not None
-
-    @property
-    def _rule_object_singular(self) -> bool:
-        return self._module_to_check_against.name is not None
-
-    def _create_rule_object_for_rule_violation_message(
-        self, rule_violations: RuleViolations
-    ) -> str:
-        object_name = self._get_module_name(self._module_to_check_against)
-        module_qualifier = "" if self._rule_object_singular else "a sub module of "
-
-        if (
-            rule_violations.should_except_violated
-            or rule_violations.should_only_except_violated_by_no_import
-        ):
-            module_qualifier = f"any module that is not {'' if self._rule_object_singular else 'a sub module of '}"
-        elif self._forbidden_import(rule_violations):
-            module_qualifier = ""
-
-            if rule_violations.strict_dependency is not None:
-                object_name = self._get_object_of_strict_dependency(rule_violations)
-
-        if (
-            rule_violations.should_not_except_violated
-            or rule_violations.should_only_violated
-        ):
-            object_name = ", ".join(
-                (module.name for module in rule_violations.lax_dependencies_found)
-            )
-
-        return f'{module_qualifier}"{object_name}"'
+        return os.sep.join(single_violation_messages)